feat: Implement comprehensive backend features for tour, booking, pass, and content management, including new services, controllers, and documentation updates.

+++ src/main/java/com/toong/modal/dto/BookingRequestDto.java
@@ -12,7 +12,7 @@ import lombok.Setter;
 @AllArgsConstructor
 public class BookingRequestDto {
 
-    private Integer departureId;
+    private Long departureId;
 
     private String firstName;
