+++ src/main/java/com/toong/repository/AdventurePassRepository.java
@@ -5,6 +5,6 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface AdventurePassRepository extends JpaRepository<AdventurePass, Integer> {
+public interface AdventurePassRepository extends JpaRepository<AdventurePass, Long> {
 
 }
