+++ src/main/java/com/toong/repository/PassOrderRepository.java
@@ -5,7 +5,7 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface PassOrderRepository extends JpaRepository<PassOrder, Integer> {
+public interface PassOrderRepository extends JpaRepository<PassOrder, Long> {
 
     PassOrder findByOrderCode(String orderCode);
